# Server/swagger.yaml
swagger: "2.0"
info:
  description: "API của Website Hỗ trợ đào tạo SGU"
  version: "1.0.0"
  title: "Website Hỗ trợ đào tạo SGU"
  contact:
    email: ""
host: ""
...
basePath: "/api"
schemes:
- "http"
- "https"
...
paths:
...
  /admin/sinh-vien/ChiTietSinhVien/{MaSV}:
    get:
      tags:
        - "SinhVien"
      summary: "Chi tiết thông tin sinh viên"
      produces:
        - application/json
      parameters:
        - in: path
          name: MaSV
          type: string
          required: true
          description: Mã sinh viên
      responses:
        200:
          description: Chi tiết thông tin sinh viên thành công.
          schema:
            type: object
            example: {
              "status": true,
              "message": "Chi tiết thông tin sinh viên thành công.",
            }
        400:
          description: Chi tiết thông tin sinh viên thất bại.
          schema:
            type: object
            example: {
              "status": false,
              "message" : "Chi tiết thông tin sinh viên thất bại.",
            }
      security:
        - ApiKeyAuth: []
...
  /admin/giang-vien/ChiTietGiangVien/{MaGV}:
    get:
      tags:
        - "GiangVien"
      summary: "Chi tiết thông tin giảng viên"
      produces:
        - application/json
      parameters:
        - in: path
          name: MaGV
          type: string
          required: true
          description: Mã giảng viên
      responses:
        200:
          description: Chi tiết thông tin giảng viên thành công.
          schema:
            type: object
            example: {
              "status": true,
              "message": "Chi tiết thông tin giảng viên thành công.",
            }
        400:
          description: Chi tiết thông tin giảng viên thất bại.
          schema:
            type: object
            example: {
              "status": false,
              "message" : "Chi tiết thông tin giảng viên thất bại.",
            }
      security:
        - ApiKeyAuth: []
...
  /sinh-vien/ChiTietSinhVien/{MaSV}:
    get:
      tags:
        - "SinhVienClient"
      summary: "Chi tiết thông tin sinh viên"
      produces:
        - application/json
      parameters:
        - in: path
          name: MaSV
          type: string
          required: true
          description: Mã sinh viên
      responses:
        200:
          description: Chi tiết thông tin sinh viên thành công.
          schema:
            type: object
            example: {
              "status": true,
              "message": "Chi tiết thông tin sinh viên thành công.",
            }
        400:
          description: Chi tiết thông tin sinh viên thất bại.
          schema:
            type: object
            example: {
              "status": false,
              "message" : "Chi tiết thông tin sinh viên thất bại.",
            }
      security:
        - ApiKeyAuth: []
  /sinh-vien/ChinhSuaThongTin/{MaSV}:
    put: 
      tags:
        - "SinhVienClient"
      summary: Chỉnh sửa thông tin sinh viên
      produces:
        - multipart/form-data       
      parameters:
        - in: path
          name: MaSV
          type: string
          required: true
          description: Mã sinh viên
        - in: formData
          name: HoSV
          type: string
          description: Họ sinh viên
          required: true
        - in: formData
          name: TenSV
          type: string
          description: Tên sinh viên
          required: true
        - in: formData
          name: Email
          type: string
          description: Email sinh viên
          required: true
        - in: formData
          name: SoDienThoai
          type: string
          description: Số điện thoại sinh viên
          required: true
        - in: formData
          name: GioiTinh
          type: string
          description: Giói tính sinh viên
          required: true
        - in: formData
          name: NgaySinh
          type: string
          description: Ngày sinh sinh viên
          required: true
      responses:
        200:
          description: Chỉnh sửa thông tin sinh viên thành công.
          schema:
            type: object
            example: {
              "status": true,
              "message": "Chỉnh sửa thông tin sinh viên thành công.",
            }          
        400:
          description: Chỉnh sửa thông tin sinh viên thất bại.
          schema:
            type: object
            example: {
              "status": false,
              "message": "Chỉnh sửa thông tin sinh viên thất bại.",
            }
      security:
        - ApiKeyAuth: [] 
  /giang-vien/ChiTietGiangVien/{MaGV}:
    get:
      tags:
        - "GiangVienClient"
      summary: "Chi tiết thông tin giảng viên"
      produces:
        - application/json
      parameters:
        - in: path
          name: MaGV
          type: string
          required: true
          description: Mã giảng viên
      responses:
        200:
          description: Chi tiết thông tin giảng viên thành công.
          schema:
            type: object
            example: {
              "status": true,
              "message": "Chi tiết thông tin giảng viên thành công.",
            }
        400:
          description: Chi tiết thông tin giảng viên thất bại.
          schema:
            type: object
            example: {
              "status": false,
              "message" : "Chi tiết thông tin giảng viên thất bại.",
            }
      security:
        - ApiKeyAuth: []
  /giang-vien/ChinhSuaThongTin/{MaGV}:
    post: 
      tags:
        - "GiangVienClient"
      summary: Chỉnh sửa thông tin giảng viên
      produces:
        - multipart/form-data       
      parameters:
        - in: path
          name: MaGV
          type: string
          required: true
          description: Mã giảng viên
        - in: formData
          name: HoGV
          type: string
          description: Họ giảng viên
          required: true
        - in: formData
          name: TenGV
          type: string
          description: Tên giảng viên
          required: true
        - in: formData
          name: Email
          type: string
          description: Email giảng viên
          required: true
        - in: formData
          name: SoDienThoai
          type: string
          description: Số điện thoại giảng viên
          required: true
        - in: formData
          name: GioiTinh
          type: string
          description: Giói tính giảng viên
          required: true
        - in: formData
          name: NgaySinh
          type: string
          description: Ngày sinh giảng viên
          required: true
        - in: formData
          name: DonViCongTac
          type: string
          description: Đơn vị công tác
          required: true
        - in: formData
          name: ChuyenNganh
          type: string
          description: Chuyên ngành của giảng viên
          required: true
        - in: formData
          name: TrinhDo
          type: string
          description: Trình độ giảng viên
          required: true
        - in: formData
          name: Hinh
          type: file
          description: Hình đại diện giảng viên
      responses:
        200:
          description: Chỉnh sửa thông tin giảng viên thành công.
          schema:
            type: object
            example: {
              "status": true,
              "message": "Chỉnh sửa thông tin giảng viên thành công.",
            }          
        400:
          description: Chỉnh sửa thông tin giảng viên thất bại.
          schema:
            type: object
            example: {
              "status": false,
              "message": "Chỉnh sửa thông tin giảng viên thất bại.",
            }
      security:
        - ApiKeyAuth: []
...
securityDefinitions:
  ApiKeyAuth:
    type: apiKey
    in: headers
    name: x-access-token